Add Panic Handling to ListUsers expandUnion
